extract seed-derivable crate
