+++ sdk/seed-derivable/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "solana-seed-derivable"
+description = "Solana trait defining the interface by which keys are derived."
+documentation = "https://docs.rs/solana-seed-derivable"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+ed25519-dalek = { workspace = true }
+ed25519-dalek-bip32 = { workspace = true }
+solana-derivation-path = { workspace = true }
+solana-signer = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
