+++ sdk/signer/src/lib.rs
@@ -2,7 +2,6 @@
 use {
     crate::presigner::PresignerError,
     itertools::Itertools,
-    solana_derivation_path::DerivationPath,
     solana_pubkey::Pubkey,
     solana_signature::Signature,
     solana_transaction_error::TransactionError,
@@ -158,20 +157,6 @@ pub trait EncodableKey: Sized {
 
         self.write(&mut f)
     }
-}
-
-/// The `SeedDerivable` trait defines the interface by which cryptographic keys/keypairs are
-/// derived from byte seeds, derivation paths, and passphrases.
-pub trait SeedDerivable: Sized {
-    fn from_seed(seed: &[u8]) -> Result<Self, Box<dyn error::Error>>;
-    fn from_seed_and_derivation_path(
-        seed: &[u8],
-        derivation_path: Option<DerivationPath>,
-    ) -> Result<Self, Box<dyn error::Error>>;
-    fn from_seed_phrase_and_passphrase(
-        seed_phrase: &str,
-        passphrase: &str,
-    ) -> Result<Self, Box<dyn error::Error>>;
 }
 
 /// The `EncodableKeypair` trait extends `EncodableKey` for asymmetric keypairs, i.e. have
