+++ sdk/src/lib.rs
@@ -94,6 +94,7 @@ pub mod rpc_port;
 pub mod secp256k1_instruction;
 pub mod shred_version;
 pub mod signature;
+pub mod signer;
 pub mod simple_vote_transaction_checker;
 pub mod system_transaction;
 pub mod transaction;
