+++ sdk/src/signer/keypair.rs
@@ -1,0 +1,1 @@
+pub use {solana_seed_derivable::keypair_from_seed_and_derivation_path, solana_signer::keypair::*};
