+++ sdk/src/signer/mod.rs
@@ -1,0 +1,9 @@
+#![cfg(feature = "full")]
+pub use {
+    solana_seed_derivable::SeedDerivable,
+    solana_signer::{
+        null_signer, presigner, signers, unique_signers, EncodableKey, EncodableKeypair, Signer,
+        SignerError,
+    },
+};
+pub mod keypair;
